// sanity.types.ts
export declare const internalGroqTypeReferenceTo: unique symbol;

// Source: schema.json
export type SanityImageAssetReference = {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
};

export type BlockContent = Array<
  | {
      children?: Array<{
        marks?: Array<string>;
        text?: string;
        _type: "span";
        _key: string;
      }>;
      style?:
        | "normal"
        | "h1"
        | "h2"
        | "h3"
        | "h4"
        | "h5"
        | "h6"
        | "blockquote"
        | "center";
      listItem?: "bullet" | "number";
      markDefs?: Array<
        | ({
            _key: string;
          } & TextColor)
        | {
            href?: string;
            _type: "link";
            _key: string;
          }
      >;
      level?: number;
      _type: "block";
      _key: string;
    }
  | {
      asset?: SanityImageAssetReference;
      media?: unknown;
      hotspot?: SanityImageHotspot;
      crop?: SanityImageCrop;
      _type: "image";
      _key: string;
    }
>;
// ...
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top: number;
  bottom: number;
  left: number;
  right: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x: number;
  y: number;
  height: number;
  width: number;
};
// ...
export type TextColor = {
  _type: "textColor";
  label?: string;
  value?: string;
};
// ...
export type HomePageQueryResult = {
  _id: "homePage";
  _type: "homePage";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  seo: {
    _type: "seo";
    seoTitle: string;
    seoDescription: string;
    seoImage: {
      asset: {
        url: string;
      };
    };
    seoKeywords: Array<string>;
  };
  herobannerLeftImages: Array<{
    asset: SanityImageAssetReference;
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
    _key: string;
  }>;
  herobannerRightImages: Array<{
    asset?: SanityImageAssetReference;
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
    _key: string;
  }>;
  herobannerHeading: BlockContent;
  herobannerTitle: BlockContent;
  herobannerDescription: BlockContent;
} | null;